reorganized hyperloglog into a separate mod
moved hyperloglog into hyperloglog/mod.rs
moved hyperloglog_data.rs into a private mod hyperloglog/data.rs
moved serde module into it's own file
added a getter for the underlying registers
